+++ .gitignore
@@ -7,4 +7,5 @@ debug.txt
 info.txt
 errors.txt
 warnings.txt
+query.txt
 dev/
