+++ lib/mysqlimport.js
@@ -13,7 +13,7 @@ const mysql = require('mysql2');
 const fs = require('fs');
 const path = require("path");
 const stream = require('stream');
-const BackLog = require('../../ClusterOperator/Backlog');
+const BackLog = require('../ClusterOperator/Backlog');
 
 
 /**
