--- modules/mysql-import/src/header.js
@@ -1,6 +1,0 @@
-'use strict';
-
-const mysql = require('mysql2');
-const fs = require('fs');
-const path = require("path");
-const stream = require('stream');
