--- query.txt
@@ -1,1 +1,0 @@
-version: 1.2.4<br>
